First attempt at creating a stats lib
